feat: apps deploy improvements (#301)
* feat: apps deploy improvements

- chore: Make dockerhub deployment generic
- fix: Feature flag error
- chore: Small refactor on build cmds
- feat: Include how to inspect logs of failed build
- fix: Avoid empty lines when not necessary
- feat: Improve error message when invalid token

before:

```
$ meroxa anything
Error: Post "https://api.meroxa.io/v1/sources": oauth2: cannot fetch token: 403 Forbidden
Response: {"error":"invalid_grant","error_description":"Unknown or invalid refresh token."}
```

after:

```
$ meroxa anything
Error: unknown or invalid refresh token, please run `meroxa login` again
```
- feat: Add spinner to logger
- feat: Include language on deploy output
- feat: Add language validation
- feat: include dashboard url

* fix: deploy js apps

* interpreting if there are functions or not

* improvements around building

* unbreak the tar.gz

* lint

* StopSpinner audit

* Pass function image name on turbine-js deploy

* fix: lint

+++ cmd/meroxa/root/apps/deploy.go
@@ -62,13 +62,15 @@ type Deploy struct {
 		DockerHubAccessToken string `long:"docker-hub-access-token" description:"DockerHub access token to use to build and deploy the app" hidden:"true"` //nolint:lll
 	}
 
-	client   deployApplicationClient
-	config   config.Config
-	logger   log.Logger
-	appName  string
-	path     string
-	lang     string
-	goDeploy turbineGo.Deploy
+	client      deployApplicationClient
+	config      config.Config
+	logger      log.Logger
+	appName     string
+	path        string
+	lang        string
+	localDeploy turbineCLI.LocalDeploy
+	fnName      string
+	tempPath    string // find something more elegant to this
 }
 
 var (
@@ -123,14 +125,14 @@ func (d *Deploy) getDockerHubAccessTokenEnv() string {
 
 func (d *Deploy) validateDockerHubFlags() error {
 	if d.flags.DockerHubUserName != "" {
-		d.goDeploy.DockerHubUserNameEnv = d.flags.DockerHubUserName
+		d.localDeploy.DockerHubUserNameEnv = d.flags.DockerHubUserName
 		if d.flags.DockerHubAccessToken == "" {
 			return errors.New("--docker-hub-access-token is required when --docker-hub-username is present")
 		}
 	}
 
 	if d.flags.DockerHubAccessToken != "" {
-		d.goDeploy.DockerHubAccessTokenEnv = d.flags.DockerHubAccessToken
+		d.localDeploy.DockerHubAccessTokenEnv = d.flags.DockerHubAccessToken
 		if d.flags.DockerHubUserName == "" {
 			return errors.New("--docker-hub-username is required when --docker-hub-access-token is present")
 		}
@@ -140,13 +142,13 @@ func (d *Deploy) validateDockerHubFlags() error {
 
 func (d *Deploy) validateDockerHubEnvVars() error {
 	if d.getDockerHubUserNameEnv() != "" {
-		d.goDeploy.DockerHubUserNameEnv = d.getDockerHubUserNameEnv()
+		d.localDeploy.DockerHubUserNameEnv = d.getDockerHubUserNameEnv()
 		if d.getDockerHubAccessTokenEnv() == "" {
 			return fmt.Errorf("%s is required when %s is present", dockerHubAccessTokenEnv, dockerHubUserNameEnv)
 		}
 	}
 	if d.getDockerHubAccessTokenEnv() != "" {
-		d.goDeploy.DockerHubAccessTokenEnv = d.getDockerHubAccessTokenEnv()
+		d.localDeploy.DockerHubAccessTokenEnv = d.getDockerHubAccessTokenEnv()
 		if d.getDockerHubUserNameEnv() == "" {
 			return fmt.Errorf("%s is required when %s is present", dockerHubUserNameEnv, dockerHubAccessTokenEnv)
 		}
@@ -168,8 +170,8 @@ func (d *Deploy) validateLocalDeploymentConfig() error {
 	}
 
 	// If there are DockerHub credentials, we consider it a local deployment
-	if d.goDeploy.DockerHubUserNameEnv != "" && d.goDeploy.DockerHubAccessTokenEnv != "" {
-		d.goDeploy.LocalDeployment = true
+	if d.localDeploy.DockerHubUserNameEnv != "" && d.localDeploy.DockerHubAccessTokenEnv != "" {
+		d.localDeploy.Enabled = true
 	}
 	return nil
 }
@@ -191,7 +193,8 @@ func (d *Deploy) createApplication(ctx context.Context, pipelineUUID, gitSha str
 		GitSha:   gitSha,
 		Pipeline: meroxa.EntityIdentifier{UUID: null.StringFrom(pipelineUUID)},
 	}
-	d.logger.Infof(ctx, "Creating application %q with language %q...", input.Name, d.lang)
+
+	d.logger.StartSpinner("\t", fmt.Sprintf("Creating application %q (%s)...", input.Name, d.lang))
 
 	res, err := d.client.CreateApplication(ctx, &input)
 	if err != nil {
@@ -200,17 +203,25 @@ func (d *Deploy) createApplication(ctx context.Context, pipelineUUID, gitSha str
 			var app *meroxa.Application
 			app, err = d.client.GetApplication(ctx, appName)
 			if err != nil {
+				d.logger.StopSpinner("\t")
 				return err
 			}
 			if app.Pipeline.UUID.String != pipelineUUID {
-				return fmt.Errorf("unable to finish creating the %s Application because its entities are in an"+
-					" unrecoverable state; try deleting and re-deploying", appName)
+				d.logger.StopSpinner(fmt.Sprintf("\t  𐄂 unable to finish creating the %s Application because its entities are in an"+
+					" unrecoverable state; try deleting and re-deploying", appName))
+
+				d.logger.StopSpinner("\t")
+				// TODO: Rollback here?
+				return fmt.Errorf("unable to finish creating application")
 			}
 		}
+		d.logger.StopSpinner("\t")
 		return err
 	}
 
-	d.logger.Infof(ctx, "Application %q successfully created!", res.Name)
+	dashboardURL := fmt.Sprintf("https://dashboard.meroxa.io/v2/apps/%s/detail", res.UUID)
+	output := fmt.Sprintf("\t✔ Application %q successfully created!\n\n  ✨ To visualize your application visit %s", res.Name, dashboardURL)
+	d.logger.StopSpinner(output)
 	d.logger.JSON(ctx, res)
 	return nil
 }
@@ -218,16 +229,23 @@ func (d *Deploy) createApplication(ctx context.Context, pipelineUUID, gitSha str
 // uploadSource creates first a Dockerfile to then, package the entire folder which will be later uploaded
 // this should ignore .git files and fixtures/.
 func (d *Deploy) uploadSource(ctx context.Context, appPath, url string) error {
-	// Before creating a .tar.zip, we make sure it contains a Dockerfile.
-	err := turbine.CreateDockerfile(appPath)
-	if err != nil {
-		return err
+	var err error
+
+	if d.lang == GoLang {
+		err = turbine.CreateDockerfile(appPath)
+		if err != nil {
+			return err
+		}
 	}
 
 	dFile := fmt.Sprintf("turbine-%s.tar.gz", d.appName)
 
 	var buf bytes.Buffer
-	d.logger.Infof(ctx, "Packaging application located at %q...", appPath)
+
+	if d.lang == JavaScript || d.lang == Python {
+		appPath = d.tempPath
+	}
+
 	err = turbineCLI.CreateTarAndZipFile(appPath, &buf)
 	if err != nil {
 		return err
@@ -248,10 +266,12 @@ func (d *Deploy) uploadSource(ctx context.Context, appPath, url string) error {
 		return err
 	}
 
-	// We clean up Dockerfile as last step
-	err = os.Remove(filepath.Join(appPath, "Dockerfile"))
-	if err != nil {
-		return err
+	if d.lang == GoLang {
+		// We clean up Dockerfile as last step
+		err = os.Remove(filepath.Join(appPath, "Dockerfile"))
+		if err != nil {
+			return err
+		}
 	}
 
 	err = d.uploadFile(ctx, dFile, url)
@@ -259,30 +279,32 @@ func (d *Deploy) uploadSource(ctx context.Context, appPath, url string) error {
 		return err
 	}
 
+	// TODO: Remove d.tempPath for JS and Python apps
 	// remove .tar.gz file
 	return os.Remove(dFile)
 }
 
 func (d *Deploy) uploadFile(ctx context.Context, filePath, url string) error {
-	d.logger.Info(ctx, "Uploading file to our build service...")
+	d.logger.StartSpinner("\t", " Uploading source...")
+
 	fh, err := os.Open(filePath)
 	if err != nil {
+		d.logger.StopSpinner("\t")
 		return err
 	}
 	defer func(fh *os.File) {
-		err = fh.Close()
-		if err != nil {
-			d.logger.Warn(ctx, err.Error())
-		}
+		fh.Close()
 	}(fh)
 
 	req, err := http.NewRequestWithContext(ctx, "PUT", url, fh)
 	if err != nil {
+		d.logger.StopSpinner("\t")
 		return err
 	}
 
 	fi, err := fh.Stat()
 	if err != nil {
+		d.logger.StopSpinner("\t")
 		return err
 	}
 
@@ -296,24 +318,30 @@ func (d *Deploy) uploadFile(ctx context.Context, filePath, url string) error {
 	client := &http.Client{}
 	res, err := client.Do(req) //nolint:bodyclose
 	if err != nil {
+		d.logger.StopSpinner("\t")
 		return err
 	}
 	defer func(Body io.ReadCloser) {
 		err := Body.Close()
-		d.logger.Infof(ctx, "Uploaded!")
 		if err != nil {
 			d.logger.Error(ctx, err.Error())
 		}
 	}(res.Body)
 
+	d.logger.StopSpinner("\t✔ Source uploaded!")
 	return nil
 }
 
 func (d *Deploy) getPlatformImage(ctx context.Context, appPath string) (string, error) {
+	d.logger.StartSpinner("\t", " Fetching Meroxa Platform source...")
+
 	s, err := d.client.CreateSource(ctx)
 	if err != nil {
+		d.logger.Errorf(ctx, "\t 𐄂 Unable to fetch source")
+		d.logger.StopSpinner("\t")
 		return "", err
 	}
+	d.logger.StopSpinner("\t✔ Platform source fetched!")
 
 	err = d.uploadSource(ctx, appPath, s.PutUrl)
 	if err != nil {
@@ -323,86 +351,164 @@ func (d *Deploy) getPlatformImage(ctx context.Context, appPath string) (string, 
 	sourceBlob := meroxa.SourceBlob{Url: s.GetUrl}
 	buildInput := &meroxa.CreateBuildInput{SourceBlob: sourceBlob}
 
+	d.logger.StartSpinner("\t", " Building Meroxa Process image...")
+
 	build, err := d.client.CreateBuild(ctx, buildInput)
 	if err != nil {
+		d.logger.StopSpinner("\t")
 		return "", err
 	}
 
-	fmt.Printf("Getting status for build: %s ", build.Uuid)
 	for {
-		fmt.Printf(".")
 		b, err := d.client.GetBuild(ctx, build.Uuid)
 		if err != nil {
+			d.logger.StopSpinner("\t")
 			return "", err
 		}
 
 		switch b.Status.State {
 		case "error":
-			return "", fmt.Errorf("build with uuid %q errored ", b.Uuid)
+			d.logger.StopSpinner(fmt.Sprintf("\t 𐄂 build with uuid %q errored\nRun `meroxa build logs %s` for more information", b.Uuid, b.Uuid))
+			return "", fmt.Errorf("build with uuid %q errored", b.Uuid)
 		case "complete":
-			fmt.Println("\nImage built! ")
+			d.logger.StopSpinner("\t✔ Process image built!")
 			return build.Image, nil
 		}
 		time.Sleep(pollDuration)
 	}
 }
 
-// Deploy takes care of all the necessary steps to deploy a Turbine application
-//	1. Build binary // different for jS
-//	2. Build image // common
-//	3. Push image // common
-//	4. Run Turbine deploy // different
-func (d *Deploy) deploy(ctx context.Context, appPath string, l log.Logger) (string, error) {
-	var fqImageName string
-	d.appName = path.Base(appPath)
+func (d *Deploy) deployApp(ctx context.Context, imageName string) (string, error) {
+	var output string
+	var err error
 
-	err := turbineGo.BuildBinary(ctx, l, appPath, d.appName, true)
+	d.logger.StartSpinner("\t", fmt.Sprintf(" Deploying application %q...", d.appName))
+	switch d.lang {
+	case GoLang:
+		output, err = turbineGo.RunDeployApp(ctx, d.logger, d.path, d.appName, imageName)
+	case JavaScript:
+		// TODO: @james
+		// TODO: Do less here!!!
+		output, err = turbineJS.Deploy(ctx, d.path, imageName, d.logger)
+	case Python:
+		// TODO: @eric
+	}
+
 	if err != nil {
+		d.logger.StopSpinner("\t𐄂 Deployment failed\n\n")
 		return "", err
 	}
 
-	var ok bool
-	// check for image instances
-	if ok, err = turbineGo.NeedsToBuild(appPath, d.appName); ok {
-		if err != nil {
-			l.Errorf(ctx, err.Error())
-			return "", err
-		}
-
-		if d.goDeploy.LocalDeployment {
-			fqImageName, err = d.goDeploy.GetDockerImageName(ctx, l, appPath, d.appName)
-			if err != nil {
-				return "", err
-			}
-		} else {
-			fqImageName, err = d.getPlatformImage(ctx, appPath)
-			if err != nil {
-				return "", err
-			}
-		}
-	}
-
-	// creates all resources
-	output, err := turbineGo.RunDeployApp(ctx, l, appPath, d.appName, fqImageName)
-	if err != nil {
-		return output, err
-	}
+	d.logger.StopSpinner("\t✔ Deploy complete!")
 	return output, nil
 }
 
-func (d *Deploy) Execute(ctx context.Context) error {
+// buildApp will call any specifics to the turbine library to prepare a directory that's ready
+// to compress, and build, to then later on call the specific command to deploy depending on the language.
+func (d *Deploy) buildApp(ctx context.Context) error {
+	var err error
+
+	d.logger.StartSpinner("\t", "Building application...")
+
+	switch d.lang {
+	case GoLang:
+		err = turbineGo.BuildBinary(ctx, d.logger, d.path, d.appName, true)
+	case JavaScript:
+		d.tempPath, err = turbineJS.BuildApp(d.path)
+	case Python:
+		// TODO: @eric
+		// path only for zipping = turbinePy.BuildApp(ctx, d.logger, d.path) => newPath string
+		// Dockerfile will already exist
+	}
+	if err != nil {
+		d.logger.StopSpinner("\t")
+		return err
+	}
+	d.logger.StopSpinner("\t✔ Application built!")
+	return nil
+}
+
+// getAppImage will check what type of build needs to perform and ultimately will return the image name to use
+// when deploying.
+func (d *Deploy) getAppImage(ctx context.Context) (string, error) {
+	d.logger.StartSpinner("\t", "Checking if application has processes...")
+	var fqImageName string
+	var needsToBuild bool
+	var err error
+
+	switch d.lang {
+	case GoLang:
+		needsToBuild, err = turbineGo.NeedsToBuild(d.path, d.appName)
+	case JavaScript:
+		needsToBuild, err = turbineJS.NeedsToBuild(d.path)
+	case Python:
+		// TODO: @eric
+	}
+	if err != nil {
+		d.logger.StopSpinner("\t")
+		return "", err
+	}
+
+	// If no need to build, return empty imageName which won't be utilized by the deploy process anyways
+	if !needsToBuild {
+		d.logger.StopSpinner("\t✔ No need to create process image...")
+		return "", nil
+	}
+
+	d.logger.StopSpinner("\t✔ Application processes found. Creating application image...")
+
+	d.localDeploy.TempPath = d.tempPath
+	d.localDeploy.Lang = d.lang
+	if d.localDeploy.Enabled {
+		fqImageName, err = d.localDeploy.GetDockerImageName(ctx, d.logger, d.path, d.appName, d.lang)
+		if err != nil {
+			return "", err
+		}
+	} else {
+		fqImageName, err = d.getPlatformImage(ctx, d.path)
+		if err != nil {
+			return "", err
+		}
+	}
+
+	return fqImageName, nil
+}
+
+// validateLanguage stops execution of the deployment in case language is not supported.
+// It also consolidates lang used in API in case user specified a supported language using an unexpected description.
+func (d *Deploy) validateLanguage() error {
+	switch d.lang {
+	case "go", GoLang:
+		d.lang = GoLang
+	case "js", JavaScript, NodeJs:
+		d.lang = JavaScript
+	case "py", Python:
+		d.lang = Python
+	default:
+		return fmt.Errorf("language %q not supported. %s", d.lang, LanguageNotSupportedError)
+	}
+	return nil
+}
+
+func (d *Deploy) validate(ctx context.Context) error {
 	// validateLocalDeploymentConfig will look for DockerHub credentials to determine whether it's a local deployment or not.
 	err := d.validateLocalDeploymentConfig()
 	if err != nil {
 		return err
 	}
-	var deployOutput string
 
 	d.path, err = turbineCLI.GetPath(d.flags.Path)
 	if err != nil {
 		return err
 	}
+	d.appName = path.Base(d.path)
+
 	d.lang, err = turbineCLI.GetLangFromAppJSON(d.path)
+	if err != nil {
+		return err
+	}
+
+	err = d.validateLanguage()
 	if err != nil {
 		return err
 	}
@@ -412,27 +518,34 @@ func (d *Deploy) Execute(ctx context.Context) error {
 		return err
 	}
 
-	err = turbineCLI.ValidateBranch(d.path)
+	return turbineCLI.ValidateBranch(ctx, d.logger, d.path)
+}
+
+func (d *Deploy) prepareAppForDeployment(ctx context.Context) error {
+	d.logger.Infof(ctx, "Preparing application %q (%s) for deployment...", d.appName, d.lang)
+
+	// After this point, CLI will package it up and will build it
+	err := d.buildApp(ctx)
 	if err != nil {
 		return err
 	}
 
-	// 1. set up the app structure (CLI does this for any language)
-	// 2. *depending on the language* call something to create the dockerfile <=
-	// 3. CLI would handle:
-	//	3.1 creating the tar.zip,
-	//	3.2 post /sources
-	// 	3.3 uploading the tar.zip
-	//  3.4 post /builds
-	// 4. CLI would call (depending on language) the deploy script <=
-	switch d.lang {
-	case GoLang:
-		deployOutput, err = d.deploy(ctx, d.path, d.logger)
-	case "js", JavaScript, NodeJs:
-		deployOutput, err = turbineJS.Deploy(ctx, d.path, d.logger)
-	default:
-		return fmt.Errorf("language %q not supported. %s", d.lang, LanguageNotSupportedError)
+	d.fnName, err = d.getAppImage(ctx)
+	return err
+}
+
+func (d *Deploy) Execute(ctx context.Context) error {
+	err := d.validate(ctx)
+	if err != nil {
+		return err
 	}
+
+	err = d.prepareAppForDeployment(ctx)
+	if err != nil {
+		return err
+	}
+
+	deployOutput, err := d.deployApp(ctx, d.fnName)
 	if err != nil {
 		return err
 	}
@@ -441,6 +554,7 @@ func (d *Deploy) Execute(ctx context.Context) error {
 	if err != nil {
 		return err
 	}
+
 	gitSha, err := turbineCLI.GetGitSha(d.path)
 	if err != nil {
 		return err
